Prisma has detected that this project was compiled on Vercel, which caches dependencies. This results in an outdated Prisma client because Prisma auto-build is not triggered. To resolve this issue, be sure to run the "prisma generate" command during the build process.
Learn more: https://pris.ly/d/vercel-build

PrismaClientInitializationError: Prisma has detected that this project was compiled on Vercel, which caches dependencies. This results in an outdated Prisma client because Prisma auto-build is not triggered. To resolve this issue, be sure to run the "prisma generate" command during the build process.